backend/requirements.txt
# Web Framework
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.1
pydantic-settings==2.9.1
python-multipart==0.0.20
python-jose==3.4.0
passlib==1.7.4
bcrypt==4.3.0
email-validator==2.2.0

# Database
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Task Queue
celery==5.5.0
redis==5.0.1

# AI & ML
spacy==3.7.2
sentence-transformers==2.2.2
lightgbm==4.1.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.2
faiss-cpu==1.7.4



# Web Scraping
scrapy==2.11.0
playwright==1.39.0

# API Clients
requests==2.31.0
httpx==0.25.0
stripe==7.3.0

# AWS
boto3==1.28.65
botocore==1.31.65

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2